requirements.txt
beautifulsoup4==4.5.3
Django==1.10.6
django-cors-headers==2.0.2
django-jet==1.0.6
django-location-field==2.0.3
django-reversion==2.0.8
djangorestframework==3.6.2
packaging==16.8
Pillow==4.0.0
pycrypto==2.6.1
pytz==2016.10
requests==2.13.0
